Upgraded go and golangci-lint (#963)
This PR upgrades GO to the latest version (1.23.2) for fixing
vulnerabilities CVE-2022-30635, CVE-2024-34156, and CVE-2024-34158.

Go 1.23.2 requires golangci-lint to be version 1.60.0 or higher, so I
upgraded golangci-lint to the latest version, 1.61.0.

+++ pkg/controllers/vdb/podfacts.go
@@ -161,7 +161,7 @@ type PodFact struct {
 	depotDiskPercentSize string
 
 	// The size of the depot in bytes.  This is only valid if the database is up.
-	maxDepotSize int
+	maxDepotSize uint64
 
 	// The size, in bytes, of the local PV.
 	localDataSize int
